+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-05-13
+
 ### Added
 
 - `group.file_browser(multiple=True)` now maps to repeated CLI options, e.g.